Improve handling of VMs in an error state.
The new behavior is to simply remove the VM and the application from the run
list. Under the old behavior we would continue to wait until the VM came back
in a good state. Since I have observed that so far all unknown states were
terminal error states, it is better to bail early.