Allow remote processing of CTRL-C.
This is great for running "ravtest run bash". If you press CTRL-C then a SIGINT
will be delivered to the remote process rather than terminating Fabric.

This is a provisional implementation that may need to be revised. First, it
depends on https://github.com/fabric/fabric/pull/823. If that pull request is
not accepted or accepted in a modified form, this implementation needs to be
updated.

Secondly, we should probably think more about unconditionally enabling this.
Possible we should detect whether the user is running a shell using some
heuristics (e.g. command == 'bash', 'csh', ...) or using a command-line option.