Always publish to Amazon.
This is in preparation for the next patch that will allow you to specify
non-default memory and CPU options. This might pick another cloud.
But because amazon is currently the only one that allows us to stop a VM
rather than destroying it, it is better to stick with Amazon for now.